final-pjt-back/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==4.2.1
django-allauth==0.52.0
django-appconf==1.0.5
django-cors-headers==4.0.0
django-imagekit==4.1.0
djangorestframework==3.14.0
idna==3.4
jellyfish==0.11.2
joblib==1.2.0
numpy==1.24.3
oauthlib==3.2.2
pilkit==2.0
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==2.0.2
